Fix amd for v5.0-rc6

+++ drivers/gpu/drm/amd/amdgpu/amdgpu_bo_list.c
@@ -70,7 +70,7 @@ int amdgpu_bo_list_create(struct amdgpu_device *adev, struct drm_file *filp,
 	unsigned i;
 	int r;
 
-	if (num_entries > (SIZE_MAX - sizeof(struct amdgpu_bo_list))
+	if ((long)num_entries > (SIZE_MAX - sizeof(struct amdgpu_bo_list))
 				/ sizeof(struct amdgpu_bo_list_entry))
 		return -EINVAL;
